feat: testable structure for api BREAKING CHANGE
