+++ src/BusinessApiTest.spec.ts
@@ -1,0 +1,9 @@
+import { expect } from "chai";
+
+describe("BusinessApiTest", () => {
+  it("Specify tests", () => {
+    expect.fail(
+      "Specify structure, behaviour and tests for BusinessApiTest mocking"
+    );
+  });
+});
