+++ src/index.ts
@@ -1,1 +1,3 @@
 export * from "./BusinessApi";
+export * from "./BusinessApiHTTP";
+export * from "./BusinessApiTest";
